Refactor Elasticsearch response handling in Storage
- Introduced a new handleResponse method to centralize response processing, including error handling and logging for Elasticsearch operations.
- Removed redundant code from existing methods by utilizing the new handleResponse function, improving maintainability and readability.
- Restored the GetIndexManager method for accessing the index manager.
